[UPD]'picking_document': Migración a v11
