+++ project-addons/picking_document/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 ##############################################################################
 #
 #    Copyright (C) 2014 Pexego Sistemas Informáticos All Rights Reserved
@@ -19,5 +18,4 @@
 #
 ##############################################################################
 
-import document
-import stock
+from . import models
