+++ project-addons/picking_document/models/__init__.py
@@ -1,0 +1,2 @@
+from . import document
+from . import stock
