+++ project-addons/picking_document/views/document_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
         <record id="document_tree" model="ir.ui.view">
@@ -49,7 +49,7 @@
             <field name="search_view_id" ref="document_filter"/>
         </record>
 
-        <menuitem name="Document" parent="stock.menu_stock_configuration" action="action_document" id="menu_document"/>
+        <menuitem name="Documents" parent="stock.menu_stock_config_settings" action="action_document" id="menu_document"/>
 
     </data>
-</openerp>
+</odoo>
