+++ project-addons/picking_document/views/stock_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
     <record id="picking_add_documents" model="ir.ui.view">
@@ -21,4 +21,4 @@
     </record>
 
     </data>
-</openerp>
+</odoo>
